requirements.txt
langchain
langchain_core
langchain_openai
langchain_postgres
python-dotenv
streamlit